Consolidate files
